thrift: split transport/protocol code into interface and implementation
Changes the organization of code, but there are no functional changes in
this PR. I'm doing this in preparation for adding more code to the transport
and protocol implementations.

*Risk Level*: low
*Testing*: existing tests
*Docs Changes*: n/a
*Release Notes*: n/a
