+++ source/extensions/filters/network/thrift_proxy/binary_protocol_impl.h
@@ -5,7 +5,7 @@
 #include "envoy/buffer/buffer.h"
 #include "envoy/common/pure.h"
 
-#include "extensions/filters/network/thrift_proxy/protocol.h"
+#include "extensions/filters/network/thrift_proxy/protocol_impl.h"
 
 namespace Envoy {
 namespace Extensions {
