+++ source/extensions/filters/network/thrift_proxy/compact_protocol_impl.h
@@ -6,7 +6,7 @@
 #include "envoy/buffer/buffer.h"
 #include "envoy/common/pure.h"
 
-#include "extensions/filters/network/thrift_proxy/protocol.h"
+#include "extensions/filters/network/thrift_proxy/protocol_impl.h"
 
 #include "absl/types/optional.h"
