+++ source/extensions/filters/network/thrift_proxy/filter.cc
@@ -5,8 +5,8 @@
 #include "common/common/assert.h"
 
 #include "extensions/filters/network/thrift_proxy/buffer_helper.h"
-#include "extensions/filters/network/thrift_proxy/protocol.h"
-#include "extensions/filters/network/thrift_proxy/transport.h"
+#include "extensions/filters/network/thrift_proxy/protocol_impl.h"
+#include "extensions/filters/network/thrift_proxy/transport_impl.h"
 
 namespace Envoy {
 namespace Extensions {
