+++ source/extensions/filters/network/thrift_proxy/transport.h
@@ -5,7 +5,6 @@
 
 #include "envoy/buffer/buffer.h"
 
-#include "common/common/fmt.h"
 #include "common/singleton/const_singleton.h"
 
 #include "absl/types/optional.h"
@@ -91,81 +90,6 @@ public:
 };
 
 typedef std::unique_ptr<Transport> TransportPtr;
-
-/*
- * TransportImplBase provides a base class for Transport implementations.
- */
-class TransportImplBase : public virtual Transport {
-public:
-  TransportImplBase(TransportCallbacks& callbacks) : callbacks_(callbacks) {}
-
-protected:
-  void onFrameStart(absl::optional<uint32_t> size) const { callbacks_.transportFrameStart(size); }
-  void onFrameComplete() const { callbacks_.transportFrameComplete(); }
-
-  TransportCallbacks& callbacks_;
-};
-
-/**
- * FramedTransportImpl implements the Thrift Framed transport.
- * See https://github.com/apache/thrift/blob/master/doc/specs/thrift-rpc.md
- */
-class FramedTransportImpl : public TransportImplBase {
-public:
-  FramedTransportImpl(TransportCallbacks& callbacks) : TransportImplBase(callbacks) {}
-
-  // Transport
-  const std::string& name() const override { return TransportNames::get().FRAMED; }
-  bool decodeFrameStart(Buffer::Instance& buffer) override;
-  bool decodeFrameEnd(Buffer::Instance& buffer) override;
-
-  static const int32_t MaxFrameSize = 0xFA0000;
-};
-
-/**
- * UnframedTransportImpl implements the Thrift Unframed transport.
- * See https://github.com/apache/thrift/blob/master/doc/specs/thrift-rpc.md
- */
-class UnframedTransportImpl : public TransportImplBase {
-public:
-  UnframedTransportImpl(TransportCallbacks& callbacks) : TransportImplBase(callbacks) {}
-
-  // Transport
-  const std::string& name() const override { return TransportNames::get().UNFRAMED; }
-  bool decodeFrameStart(Buffer::Instance&) override {
-    onFrameStart(absl::optional<uint32_t>());
-    return true;
-  }
-  bool decodeFrameEnd(Buffer::Instance&) override {
-    onFrameComplete();
-    return true;
-  }
-};
-
-/**
- * AutoTransportImpl implements Transport and attempts to distinguish between the Thrift framed and
- * unframed transports. Once the transport is detected, subsequent operations are delegated to the
- * appropriate implementation.
- */
-class AutoTransportImpl : public TransportImplBase {
-public:
-  AutoTransportImpl(TransportCallbacks& callbacks)
-      : TransportImplBase(callbacks), name_(TransportNames::get().AUTO){};
-
-  // Transport
-  const std::string& name() const override { return name_; }
-  bool decodeFrameStart(Buffer::Instance& buffer) override;
-  bool decodeFrameEnd(Buffer::Instance& buffer) override;
-
-private:
-  void setTransport(TransportPtr&& transport) {
-    transport_ = std::move(transport);
-    name_ = fmt::format("{}({})", transport_->name(), TransportNames::get().AUTO);
-  }
-
-  TransportPtr transport_{};
-  std::string name_;
-};
 
 } // namespace ThriftProxy
 } // namespace NetworkFilters
